Renamed Select and Filter VMs with Form extension
