+++ src/app/Plans/components/Wizard/PlanWizard.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { Modal, Wizard } from '@patternfly/react-core';
 import WizardStepContainer from './WizardStepContainer';
 import GeneralForm from './GeneralForm';
-import FilterVMs from './FilterVMs';
-import SelectVMs from './SelectVMs';
+import FilterVMs from './FilterVMsForm';
+import SelectVMs from './SelectVMsForm';
 import Review from './Review';
 import { Provider } from '@app/Providers/types';
 import { MOCK_VMS } from './mocks/VMs.mock';
